+++ lib/fdevs_fitkit.dart
@@ -1,1 +1,3 @@
 library fdevs_fitkit;
+
+export './src/rendering/fitkit.dart';