+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1+1
 
 * Package Name Changed from `fitkit` to `fdevs_fitkit`
+
+## 0.0.1+2
+
+* Responsive extensions added for numbers
+* Widget extensions added with Responsive values